backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
python-socketio
python-multipart
redis
joblib
numpy
